requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
openai>=1.0.0
python-dotenv>=1.0.0
rapidfuzz>=3.6.1
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
